+++ test/ngdart_test.dart
@@ -13,7 +13,7 @@ import 'package:path/path.dart' as path;
 import 'package:test/test.dart';
 
 void main() {
-  group('ngdart', () {
+  group('ngflutter', () {
     AppLoggerMock logger;
     FileWriterMock writer;
     NgDartCommanderRunner runner;